test: unicode collections and separate pytest fixtures
