+++ tests/fixtures/api/card/30.json
@@ -2,7 +2,7 @@
     "cache_invalidated_at": null,
     "description": "Dummy 1",
     "archived": false,
-    "view_count": 43,
+    "view_count": 47,
     "collection_position": null,
     "table_id": null,
     "can_run_adhoc_query": true,
@@ -41,7 +41,7 @@
     "creator": {
         "email": "dbtmetabase@example.com",
         "first_name": "dbtmetabase",
-        "last_login": "2024-11-02T04:41:42.402168Z",
+        "last_login": "2024-12-20T12:30:47.814008Z",
         "is_qbnewb": false,
         "is_superuser": true,
         "id": 1,
