+++ tests/fixtures/api/card/33.json
@@ -2,7 +2,7 @@
     "cache_invalidated_at": null,
     "description": null,
     "archived": false,
-    "view_count": 15,
+    "view_count": 19,
     "collection_position": null,
     "table_id": 18,
     "can_run_adhoc_query": true,
@@ -146,7 +146,7 @@
     "creator": {
         "email": "dbtmetabase@example.com",
         "first_name": "dbtmetabase",
-        "last_login": "2024-11-02T04:41:42.402168Z",
+        "last_login": "2024-12-20T12:30:47.814008Z",
         "is_qbnewb": false,
         "is_superuser": true,
         "id": 1,
@@ -158,17 +158,17 @@
     "can_write": true,
     "database_id": 2,
     "enable_embedding": false,
-    "collection_id": null,
+    "collection_id": 3,
     "query_type": "query",
     "name": "Transactions",
-    "last_query_start": "2024-10-12T03:53:32.250619Z",
+    "last_query_start": "2024-12-20T12:30:05.862314Z",
     "dashboard_count": 0,
-    "last_used_at": "2024-10-12T03:43:09.869884Z",
+    "last_used_at": "2024-12-20T12:30:05.932096Z",
     "type": "question",
-    "average_query_time": 114.5,
+    "average_query_time": 134.33333333333334,
     "creator_id": 1,
     "moderation_reviews": [],
-    "updated_at": "2024-10-12T01:52:30.606845Z",
+    "updated_at": "2024-12-20T12:30:11.735631Z",
     "made_public_by_id": null,
     "embedding_params": null,
     "cache_ttl": null,
@@ -189,18 +189,26 @@
         "email": "dbtmetabase@example.com",
         "first_name": "dbtmetabase",
         "last_name": null,
-        "timestamp": "2024-10-11T23:46:29.299411Z"
+        "timestamp": "2024-12-20T12:30:11.782155Z"
     },
     "visualization_settings": {
         "table.cell_column": "payment_id"
     },
     "collection": {
-        "metabase.models.collection.root/is-root?": true,
         "authority_level": null,
-        "name": "Our analytics",
+        "description": null,
+        "archived": false,
+        "slug": "%D0%BA%D0%BE%D0%BB%D0%BB%D0%B5%D0%BA%D1%86%D0%B8%D1%8F",
+        "name": "\u041a\u043e\u043b\u043b\u0435\u043a\u0446\u0438\u044f",
+        "personal_owner_id": null,
+        "type": null,
+        "is_sample": false,
+        "id": 3,
+        "entity_id": "4Zdu-bk8RhwuFiEiRXuOH",
+        "location": "/",
+        "namespace": null,
         "is_personal": false,
-        "id": "root",
-        "can_write": true
+        "created_at": "2024-12-20T12:29:39.063908Z"
     },
     "metabase_version": "v0.50.5 (48f6978)",
     "parameters": [],
