+++ tests/fixtures/api/collection.json
@@ -25,5 +25,22 @@
         "namespace": null,
         "is_personal": true,
         "created_at": "2024-06-19T11:49:38.03895Z"
+    },
+    {
+        "authority_level": null,
+        "description": null,
+        "archived": false,
+        "slug": "%D0%BA%D0%BE%D0%BB%D0%BB%D0%B5%D0%BA%D1%86%D0%B8%D1%8F",
+        "can_write": true,
+        "name": "\u041a\u043e\u043b\u043b\u0435\u043a\u0446\u0438\u044f",
+        "personal_owner_id": null,
+        "type": null,
+        "is_sample": false,
+        "id": 3,
+        "entity_id": "4Zdu-bk8RhwuFiEiRXuOH",
+        "location": "/",
+        "namespace": null,
+        "is_personal": false,
+        "created_at": "2024-12-20T12:29:39.063908Z"
     }
 ]
