+++ tests/fixtures/api/database.json
@@ -55,7 +55,7 @@
         "caveats": null,
         "creator_id": 1,
         "is_full_sync": true,
-        "updated_at": "2024-11-02T04:30:07.348325Z",
+        "updated_at": "2024-12-20T12:28:30.976262Z",
         "native_permissions": "write",
         "cache_ttl": null,
         "details": {
