+++ tests/fixtures/api/database/2/metadata.json
@@ -412,7 +412,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 12,
             "db_id": 2,
@@ -941,7 +941,7 @@
             ],
             "caveats": "Some facts are derived from payments",
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 10,
             "db_id": 2,
@@ -1103,7 +1103,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 13,
             "db_id": 2,
@@ -1313,7 +1313,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 14,
             "db_id": 2,
@@ -1527,7 +1527,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 11,
             "db_id": 2,
@@ -1639,7 +1639,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 19,
             "db_id": 2,
@@ -2126,7 +2126,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 15,
             "db_id": 2,
@@ -2355,7 +2355,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 9,
             "db_id": 2,
@@ -2633,7 +2633,7 @@
             ],
             "caveats": null,
             "segments": [],
-            "updated_at": "2024-11-02T04:33:23.804648Z",
+            "updated_at": "2024-12-20T12:28:32.299132Z",
             "active": true,
             "id": 18,
             "db_id": 2,
@@ -2650,7 +2650,7 @@
     ],
     "creator_id": 1,
     "is_full_sync": true,
-    "updated_at": "2024-11-02T04:30:07.348325Z",
+    "updated_at": "2024-12-20T12:28:30.976262Z",
     "cache_ttl": null,
     "details": {
         "ssl": false,
