+++ tests/fixtures/api/table.json
@@ -61,7 +61,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -96,7 +96,7 @@
         "show_in_getting_started": false,
         "name": "customers",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 12,
         "db_id": 2,
@@ -171,7 +171,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -206,7 +206,7 @@
         "show_in_getting_started": false,
         "name": "orders",
         "caveats": "Some facts are derived from payments",
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 10,
         "db_id": 2,
@@ -281,7 +281,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -316,7 +316,7 @@
         "show_in_getting_started": false,
         "name": "raw_customers",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 13,
         "db_id": 2,
@@ -391,7 +391,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -426,7 +426,7 @@
         "show_in_getting_started": false,
         "name": "raw_orders",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 14,
         "db_id": 2,
@@ -501,7 +501,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -536,7 +536,7 @@
         "show_in_getting_started": false,
         "name": "raw_payments",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 11,
         "db_id": 2,
@@ -611,7 +611,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -646,7 +646,7 @@
         "show_in_getting_started": false,
         "name": "skus",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 19,
         "db_id": 2,
@@ -721,7 +721,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -831,7 +831,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -866,7 +866,7 @@
         "show_in_getting_started": false,
         "name": "stg_orders",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 15,
         "db_id": 2,
@@ -941,7 +941,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -976,7 +976,7 @@
         "show_in_getting_started": false,
         "name": "stg_payments",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 9,
         "db_id": 2,
@@ -1051,7 +1051,7 @@
             "caveats": null,
             "creator_id": 1,
             "is_full_sync": true,
-            "updated_at": "2024-11-02T04:30:07.348325Z",
+            "updated_at": "2024-12-20T12:28:30.976262Z",
             "cache_ttl": null,
             "details": {
                 "ssl": false,
@@ -1086,7 +1086,7 @@
         "show_in_getting_started": false,
         "name": "transactions",
         "caveats": null,
-        "updated_at": "2024-11-02T04:33:23.804648Z",
+        "updated_at": "2024-12-20T12:28:32.299132Z",
         "active": true,
         "id": 18,
         "db_id": 2,
