+++ tests/fixtures/api/user/1.json
@@ -2,7 +2,7 @@
     "email": "dbtmetabase@example.com",
     "first_name": "dbtmetabase",
     "locale": null,
-    "last_login": "2024-11-02T04:41:42.402168Z",
+    "last_login": "2024-12-20T12:30:47.814008Z",
     "is_active": true,
     "user_group_memberships": [
         {
@@ -13,7 +13,7 @@
         }
     ],
     "is_qbnewb": false,
-    "updated_at": "2024-11-02T04:41:42.402168Z",
+    "updated_at": "2024-12-20T12:30:47.814008Z",
     "is_superuser": true,
     "login_attributes": null,
     "id": 1,
