+++ tests/fixtures/exposure/collection/our_analytics.yml
@@ -184,29 +184,3 @@ exposures:
       - ref('orders')
       - ref('stg_orders')
       - ref('stg_payments')
-  - name: transactions
-    label: Transactions
-    description: '### Visualization: Table
-
-
-      No description provided in Metabase
-
-
-      #### Metadata
-
-
-      Metabase ID: __33__
-
-
-      Created On: __2024-10-11T23:46:29.272117Z__'
-    type: analysis
-    url: http://localhost:3000/card/33
-    maturity: medium
-    owner:
-      name: dbtmetabase
-      email: dbtmetabase@example.com
-    depends_on:
-      - ref('payments')
-    meta:
-      average_query_time: '0:00.115'
-      last_used_at: '2024-10-12T03:43:09.869884Z'
