+++ tests/fixtures/exposure/collection/коллекция.yml
