+++ tests/fixtures/exposure/default/exposures.yml
@@ -222,7 +222,7 @@ exposures:
     depends_on:
       - ref('payments')
     meta:
-      average_query_time: '0:00.115'
-      last_used_at: '2024-10-12T03:43:09.869884Z'
+      average_query_time: '0:00.134'
+      last_used_at: '2024-12-20T12:30:05.932096Z'
     tags:
       - metabase
